requirements.txt
PySide6
pytesseract
pyclip
pillow